requirements.txt
Babel==2.7.0
Django==2.2.7
django-phonenumber-field==3.0.1
mysqlclient==1.4.5
phonenumbers==8.10.22
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
